extract utils.ts, lift hooks /utils => /hooks
