+++ src/utils/preventDefault.ts
@@ -1,0 +1,3 @@
+export const preventDefault = (e: React.TouchEvent | React.MouseEvent) => {
+  e.preventDefault();
+};
